+++ CHANGELOG.md
@@ -1,13 +1,13 @@
 # Changelog
 
-All notable changes to `:package_name` will be documented in this file.
+All notable changes to `Calculator` will be documented in this file.
 
 Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) principles.
 
-## NEXT - YYYY-MM-DD
+## 0.0.1 - 2019-03-01
 
 ### Added
-- Nothing
+- Skeleton library
 
 ### Deprecated
 - Nothing